+++ pkg/core/blockchainer.go
@@ -1,6 +1,7 @@
 package core
 
 import (
+	"github.com/CityOfZion/neo-go/config"
 	"github.com/CityOfZion/neo-go/pkg/core/transaction"
 	"github.com/CityOfZion/neo-go/pkg/util"
 )
@@ -8,6 +9,7 @@ import (
 // Blockchainer is an interface that abstract the implementation
 // of the blockchain.
 type Blockchainer interface {
+	GetConfig() config.ProtocolConfiguration
 	AddHeaders(...*Header) error
 	AddBlock(*Block) error
 	BlockHeight() uint32
@@ -22,4 +24,8 @@ type Blockchainer interface {
 	GetAssetState(util.Uint256) *AssetState
 	GetAccountState(util.Uint160) *AccountState
 	GetTransaction(util.Uint256) (*transaction.Transaction, uint32, error)
+	References(t *transaction.Transaction) map[util.Uint256]*transaction.Output
+	FeePerByte(t *transaction.Transaction) util.Fixed8
+	SystemFee(t *transaction.Transaction) util.Fixed8
+	NetworkFee(t *transaction.Transaction) util.Fixed8
 }
