+++ pkg/rpc/wrappers/tx_raw_output.go
@@ -31,8 +31,8 @@ func NewTransactionOutputRaw(tx *transaction.Transaction, header *core.Header, c
 		tx,
 		tx.Hash(),
 		tx.Size(),
-		core.SystemFee(tx),
-		core.NetworkFee(tx),
+		chain.SystemFee(tx),
+		chain.NetworkFee(tx),
 		header.Hash(),
 		confirmations,
 		header.Timestamp,
